Enable warnings, fix some warnings.

+++ include/image_processing/MotionDetection.h
@@ -67,7 +67,7 @@ public:
 #endif /* CV_MAJOR_VERSION 2 */
 
     bool detect_on_cloud(const PointCloudXYZ::Ptr sv, const std::vector<double>& sv_center, PointCloudXYZ::Ptr diff_cloud,
-                         int threshold = 0,double dist_thres = 0.02, double mean_thres = 0.2, double octree_res = 0.02);
+                         size_t threshold = 0,double dist_thres = 0.02, double mean_thres = 0.2, double octree_res = 0.02);
 
     /**
      * @brief setInputFrames
